+++ .eslintrc.js
@@ -6,6 +6,7 @@ module.exports = {
     browser: true,
     es6: true,
     node: true,
+    mocha: true
   },
   parser: "@typescript-eslint/parser",
   parserOptions: {
