+++ package.json
@@ -5,7 +5,9 @@
     "main": "lib/index.js",
     "types": "lib/index.d.ts",
     "scripts": {
-        "test": "mocha -r ts-node/register test/**/*.test.ts ",
+        "test": "yarn run test:node && yarn run test:web",
+        "test:node": "mocha -r ts-node/register test/**/*.spec.ts ",
+        "test:web": "karma start",
         "clean": "rm -rf lib",
         "lint": "eslint --ext .ts src/",
         "check-types": "tsc --noEmit",
@@ -37,18 +39,20 @@
         "@typescript-eslint/eslint-plugin": "^1.3.0",
         "@typescript-eslint/parser": "^1.3.0",
         "chai": "^4.2.0",
-        "codecov": "^3.1.0",
         "eslint": "5.16.0",
         "eslint-plugin-import": "^2.18.2",
+        "karma": "^4.4.1",
+        "karma-chai": "^0.1.0",
+        "karma-chrome-launcher": "^3.1.0",
+        "karma-cli": "^2.0.0",
+        "karma-mocha": "^1.3.0",
+        "karma-spec-reporter": "^0.0.32",
+        "karma-webpack": "^4.0.2",
         "mocha": "^6.1.4",
-        "nyc": "^14.1.1",
+        "ts-loader": "^6.2.1",
         "ts-node": "^8.3.0",
-        "typedoc": "^0.15.0",
-        "typedoc-plugin-external-module-name": "^2.1.0",
-        "typedoc-plugin-internal-external": "^2.0.2",
-        "typedoc-plugin-markdown": "^2.0.6",
         "typescript": "^3.5.1",
-        "webpack": "^4.39.1"
+        "webpack": "^4.41.2"
     },
     "dependencies": {
         "assert": "^2.0.0",
